requirements.txt
Django==1.11.16
psycopg2-binary
xlwt
xlrd
djangorestframework
django-json-field
ConfigParser
urllib3
pyvmomi
pexpect
pycryptodome
ansible==2.7.10
dwebsocket